Подготовка к компилированию
плюс к этому подчистка говнокода и добавление комментариев чтобы ну прям совсем дурачок додумался где что работает

+++ client.py
@@ -1,32 +1,94 @@
 """
 Работа с Telegram через кастомное приложение. Требует регистрации в [my.telegram.org](https://my.telegram.org).
 
-Используется библиотека [Telethon](https://docs.telethon.dev/).
+Используются библиотеки [Telethon](https://docs.telethon.dev/) и [qrcode](https://github.com/lincolnloop/python-qrcode).
 """
 import logging
+from config import config
+from utils import clear
 from io import BytesIO
-from telethon import TelegramClient
+from qrcode import QRCode
+from getpass import getpass
+from telethon import TelegramClient, connection
+from telethon.errors import SessionPasswordNeededError as SPNError
 #from telethon.tl.functions.users import GetSavedMusicRequest
 from telethon.tl.functions.account import SaveMusicRequest
+from colorama import init, Fore, Back, Style
 
-
-from base64 import b64decode
+init(autoreset=True)
 
 
 logging.basicConfig(
     format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO
 )
 
-api_id = 0
-api_hash = 'nuh-uh, ты должен ввести свои!'
+config = config['telegram']
 
-client = TelegramClient('trp', api_id, api_hash)
+API_ID = config['api_id']
+API_HASH = config['api_hash']
+
+if config['proxy']['enabled']:
+    client = TelegramClient('trp', API_ID, API_HASH, connection=connection.ConnectionTcpMTProxyRandomizedIntermediate, proxy=(config['proxy']['host'], config['proxy']['port'], config['proxy']['secret']), connection_retries=-1)
+else:
+    client = TelegramClient('trp', API_ID, API_HASH)
 file = None # Чтобы не засорять чат, будет изменяться один файл и впоследствии удаляться
+qr = QRCode()
+
+def gen_qr(url):
+    """
+    Генерация QR-кода для авторизации.
+    """
+    clear()
+    print(f"{Fore.LIGHTBLUE_EX}Отсканируйте данный QR-код через приложение (Настройки - Устройства - Подключить устройство):")
+    qr.clear()
+    qr.add_data(url)
+    qr.print_ascii()
 
 async def start_client():
-    await client.start()
+    """
+    Инициализация Telegram-клиента.
+    
+    Если нет активной сессии, инициализирует интерактивную авторизацию, в ином случае просто логинит.
+    """
+    await client.connect()
+    if await client.is_user_authorized():
+        return
+    
+    choice = input(f"{Fore.LIGHTYELLOW_EX}Выберите метод аутентификации:{Fore.WHITE}\n1) По QR-коду {Fore.LIGHTBLUE_EX}(по умолчанию){Fore.WHITE}\n2) По номеру телефона\n[{Style.BRIGHT}1{Style.RESET_ALL}/2]: ")
+    while choice not in ["1", "2", ""]:
+        choice = input(f"[{Style.BRIGHT}1{Style.RESET_ALL}/2]: ")
+    if choice == "":
+        choice = "1"
+    choice = int(choice)
+    if choice == 2:
+        await client.start() # по номеру телефона
+        return
+    
+    qr_login = await client.qr_login() # по qr-коду
+    r = False
+    while not r:
+        gen_qr(qr_login.url)
+        try:
+            r = await qr_login.wait()
+        except TimeoutError:
+            await qr_login.recreate()
+        except SPNError:
+            clear()
+            tfa = False
+            r = True
+            while tfa == False:
+                pwd = getpass(f"{Fore.MAGENTA}Введите код двухфакторной аутентификации: ")
+                try:
+                    await client.sign_in(password=pwd)
+                    tfa = True
+                except Exception as e:
+                    print(f"{Fore.RED}{Style.BRIGHT}Указан неверный пароль.{Style.NORMAL} {Fore.YELLOW}Повторите попытку.\n{Fore.RED}{e}\n")
+    clear()
 
 async def stop_client():
+    """
+    Остановка клиента. Наверное.
+    """
     await client.disconnect()
 
 async def update_status(sample: BytesIO, thumb = None):
@@ -34,10 +96,6 @@ async def update_status(sample: BytesIO, thumb = None):
     Загрузка статуса и его обновление в профиле.
     """
     global file
-    _header, b64 = thumb.split(",", 1)
-    data = b64decode(b64)
-    thumb = BytesIO(data)
-    thumb.seek(0)
     if file:
         await client(SaveMusicRequest(file, unsave=True))
         file = await file.edit(file=sample, thumb=thumb)
@@ -50,17 +108,28 @@ async def delete_status():
     """
     Удаление статуса из профиля.
 
-    Всегда выдаёт True, даже если статуса для удаления нет.
+    Запросу без разницы, есть ли файл в списке музыки или его там нет.
+    
+    Другое дело, когда файла вовсе нет...
     """
     global file
-    result = await client(SaveMusicRequest(file, unsave=True))
-    return result
+    print(Fore.BLUE + 'Удаление статуса...')
+    if file is None:
+        print(Fore.YELLOW + 'Удалять нечего, пропускаем')
+        return
+    await client(SaveMusicRequest(file, unsave=True))
+    print(Fore.GREEN + 'Статус удалён')
 
 async def delete_cache():
     """
-    Удаление файла из избранных.
+    Удаление файла из избранных (если есть что удалять).
     
     Вы же не будете его там вечно хранить, так ведь?
     """
     global file
+    print(Fore.BLUE + 'Удаление файла...')
+    if file is None:
+        print(Fore.YELLOW + 'Удалять нечего, пропускаем')
+        return
     await file.delete()
+    print(Fore.GREEN + 'Файл удалён')
